+++ src/prompt/data.py
@@ -2,6 +2,7 @@ import os
 import json
 import numpy as np
 import cv2
+import time
 from PIL import Image
 
 from torch.utils.data import Dataset, DataLoader, SubsetRandomSampler, IterableDataset, get_worker_info
@@ -21,70 +22,93 @@ def get_classname():
     return convert
 
 
-class Image2TextDataset(Dataset):
-    def __init__(self, root_path, json_path, image_transform, mode='train'):
+class I2TTrainDataset(Dataset):
+    def __init__(self, root_path, json_path, image_transform):
         self.root_path = root_path
         self.dataset = json.load(open(json_path,'r'))
         self.convert = get_classname()
         self.image_transform = image_transform
-        self.mode = mode
     
     def __len__(self):
         return len(self.dataset)
     
     def __getitem__(self, index):
         long_caption = self.dataset[index]['caption']
-
-        if self.mode == 'train':
-            original_image_path = os.path.join(self.root_path, self.dataset[index]['image_path'])
-            original_image = self.image_transform(Image.open(original_image_path).convert('RGB'))
-            negative_image_path = os.path.join(self.root_path, self.dataset[np.random.randint(1, len(self.dataset))]['image_path'])
-            negative_image = self.image_transform(Image.open(negative_image_path).convert('RGB'))
-            return [original_image, long_caption, negative_image]
-        else:
-            original_image_path = os.path.join(self.root_path, self.dataset[index]['image_path'])
-            original_image = self.image_transform(Image.open(original_image_path).convert('RGB'))
-            return [original_image, long_caption]
+        original_image_path = os.path.join(self.root_path, self.dataset[index]['image_path'])
+        original_image = self.image_transform(Image.open(original_image_path).convert('RGB'))
+        negative_image_path = os.path.join(self.root_path, self.dataset[np.random.randint(1, len(self.dataset))]['image_path'])
+        negative_image = self.image_transform(Image.open(negative_image_path).convert('RGB'))
+        return [original_image, long_caption, negative_image]
 
 
-class Image2ImageDataset(Dataset):
-    def __init__(self, root_path, other_path, json_path, image_transform, mode='train'):
+class I2TTestDataset(Dataset):
+    def __init__(self, root_path, json_path, image_transform):
         self.root_path = root_path
-        self.other_path = other_path
         self.dataset = json.load(open(json_path,'r'))
         self.convert = get_classname()
         self.image_transform = image_transform
-        self.mode = mode
     
     def __len__(self):
         return len(self.dataset)
     
     def __getitem__(self, index):
+        long_caption = self.dataset[index]['caption']
+        original_image_path = os.path.join(self.root_path, self.dataset[index]['image_path'])
+        original_image = self.image_transform(Image.open(original_image_path).convert('RGB'))
+        return [original_image, long_caption]
 
-        if self.mode == 'train':
-            original_image_path = os.path.join(self.root_path, self.dataset[index]['origin_image'])
-            original_image = self.image_transform(Image.open(original_image_path).convert('RGB'))
-            classname = os.path.join(self.root_path, self.dataset[index]['classname'])
-            
-            other_image_path = os.path.join(self.other_path, self.dataset[index]['sketch_image'])
-            other_image = self.image_transform(Image.open(other_image_path).convert('RGB'))
-            
+
+class I2ITrainDataset(Dataset):
+    def __init__(self, root_path, other_path, json_path, image_transform):
+        self.root_path = root_path
+        self.other_path = other_path
+        self.dataset = json.load(open(json_path,'r'))
+        self.convert = get_classname()
+        self.image_transform = image_transform
+    
+    def __len__(self):
+        return len(self.dataset)
+    
+    def __getitem__(self, index):
+        classname = os.path.join(self.root_path, self.dataset[index]['classname'])
+
+        original_image_path = os.path.join(self.root_path, self.dataset[index]['origin_image'])
+        other_image_path = os.path.join(self.other_path, self.dataset[index]['sketch_image'])
+        
+        num = np.random.randint(1, len(self.dataset))
+        negative_class = os.path.join(self.other_path, self.dataset[num]['classname'])
+        while negative_class == classname:
             num = np.random.randint(1, len(self.dataset))
             negative_class = os.path.join(self.other_path, self.dataset[num]['classname'])
-            while negative_class == classname:
-                num = np.random.randint(1, len(self.dataset))
-                negative_class = os.path.join(self.other_path, self.dataset[num]['classname'])
-            
-            negative_image_path = os.path.join(self.other_path, self.dataset[num]['sketch_image'])
-            negative_image = self.image_transform(Image.open(negative_image_path).convert('RGB'))
+        negative_image_path = os.path.join(self.other_path, self.dataset[num]['sketch_image'])
 
-            return [original_image, other_image, negative_image]
-        else:
-            original_image_path = os.path.join(self.root_path, self.dataset[index]['sketch_image'])
-            original_image = self.image_transform(Image.open(original_image_path).convert('RGB'))
-            other_image_path = os.path.join(self.other_path, self.dataset[index]['image_path'])
-            other_image = self.image_transform(Image.open(other_image_path).convert('RGB'))
-            return [original_image, other_image]
+        original_image = self.image_transform(Image.open(original_image_path))
+        other_image = self.image_transform(Image.open(other_image_path))
+        negative_image = self.image_transform(Image.open(negative_image_path))
+
+        return [original_image, other_image, negative_image]
+
+
+class I2ITestDataset(Dataset):
+    def __init__(self, root_path, other_path, root_json_path, other_json_path, image_transform):
+        self.root_path = root_path
+        self.other_path = other_path
+        self.root_dataset = json.load(open(root_json_path,'r'))
+        self.other_dataset = json.load(open(other_json_path,'r'))
+        self.convert = get_classname()
+        self.image_transform = image_transform
+    
+    def __len__(self):
+        return len(self.other_dataset)
+    
+    def __getitem__(self, index):
+        original_image_path = os.path.join(self.root_path, self.root_dataset[index]['image'])
+        original_image = self.image_transform(Image.open(original_image_path))
+        original_classname = self.root_dataset[index]['classname']
+        other_image_path = os.path.join(self.other_path, self.other_dataset[index]['image'])
+        other_image = self.image_transform(Image.open(other_image_path))
+        other_image_classname = self.other_dataset[index]['classname']
+        return [original_image, other_image, original_classname, other_image_classname]
         
 
 class DataLoaderX(DataLoader):
@@ -92,7 +116,7 @@ class DataLoaderX(DataLoader):
         return BackgroundGenerator(super().__iter__())
     
 
-class data_prefetcher():
+class DataPrefetcher():
     def __init__(self, loader):
         #loader 1：real
         #loader 2：fake
@@ -100,21 +124,22 @@ class data_prefetcher():
         self.loader = iter(loader)
         self.preload()
  
- 
+
     def preload(self):
         try:
-            self.oimage = next(self.loader)
+            self.data = next(self.loader)
         except StopIteration:
-            self.oimage = None
+            self.data = None
             return
         with torch.cuda.stream(self.stream):
-            self.oimage = [self.oimage[0].cuda(non_blocking=True).float(),
-                        self.oimage[1].cuda(non_blocking=True).float(),
-                        self.oimage[2].cuda(non_blocking=True).float()]
+            temp = []
+            for item in range(len(self.data)):
+                temp.append(self.data[item].cuda(non_blocking=True).float())
+            self.data = temp
  
 
     def next(self):
         torch.cuda.current_stream().wait_stream(self.stream)
-        oimage = self.oimage
+        data = self.data
         self.preload()
-        return oimage
+        return data